requirements.txt
# local package
-e .

# external requirements
autocorrect
click
Sphinx
coverage
awscli
flake8
pandas==1.2.1
python-dotenv>=0.5.1
validators
spam-lists
re
pyspf
checkdmarc
spam_lists
dnspython